model/requirements.txt
flask
flask-cors
ultralytics
opencv-python-headless
numpy
rasterio
pyproj
geopandas
python-dotenv
torch
torchvision
ensemble-boxes
gunicorn